Function libimaginteraction::ask::ask_string

pub fn ask_string(s: &str, default: Option<String>, permit_empty: bool, permit_multiline: bool, eof: Option<&str>, prompt: &str) -> String

Ask the user for a String.

If permit_empty is set to false, the default value will be returned if the user inserts an empty string.

If the permit_empty value is true, the default value is never returned.

If the permit_multiline is set to true, the prompt will be displayed before each input line.

If the eof parameter is None, the input ends as soon as there is an empty line input from the user. If the parameter is Some(text), the input ends if the input line is equal to text.